+++ test/configuration_test.rb
@@ -30,7 +30,7 @@ class ConfigurationTest < Minitest::Test
     assert_equal Webpacker::Configuration.source_path.to_s, source_path
   end
 
-  def test_compile_missing_packs
-    assert Webpacker::Configuration.compile_missing_packs?
+  def test_compile?
+    refute Webpacker::Configuration.compile?
   end
 end